Move fetch tags from repo cards to tag side bar
